requirements.txt
streamlit>=1.36
openai>=1.35
pydantic>=2.7
PyYAML>=6.0.1

